# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Insurance Pricing API",
    "summary": "HTTP inference API for the insurance premium pricing model.",
    "description": "Production-style FastAPI surface for serving insurance pricing predictions.\n\n## Scope\n- score raw insurance business records\n- expose frequency, severity, and final premium predictions\n- keep the served model bundle pinned at startup through `INSURANCE_PRICING_RUN_ID`\n\n## Request Contract\nSend the raw columns from the historical scoring dataset, such as the ones present in `data/test.csv`. Feature engineering stays inside the service layer, so clients only need to send business-facing fields.\n\n## Response Headers\nEvery response includes `X-Request-ID` and `X-Process-Time-Ms` headers to help trace calls across clients, logs, and deployment environments.\n\n## Batch Behavior\nBatch endpoints preserve input order and return the optional `index` field when it is provided by the caller.\n\n## Product Quote Endpoints\nThe public `/quotes` endpoints persist end-user quote history scoped to an anonymous `X-Client-ID` header so the Next.js frontend can retrieve prior simulations.",
    "contact": {
      "name": "Insurance Pricing Team"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/quotes": {
      "post": {
        "tags": [
          "quotes"
        ],
        "summary": "Create a persisted quote",
        "description": "Scores one insurance record, persists the input payload and the resulting quote, and returns the stored quote detail for the current customer context.",
        "operationId": "create_quote",
        "parameters": [
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionInput",
                "description": "Raw business-facing fields for one insurance contract. The payload mirrors the technical prediction contract used by the model."
              },
              "examples": {
                "single_record": {
                  "summary": "Single contract quote example",
                  "value": {
                    "index": 50000,
                    "bonus": 0.58,
                    "type_contrat": "Maxi",
                    "duree_contrat": 1,
                    "anciennete_info": 1,
                    "freq_paiement": "Yearly",
                    "paiement": "No",
                    "utilisation": "Retired",
                    "code_postal": 28388,
                    "conducteur2": "No",
                    "age_conducteur1": 66,
                    "age_conducteur2": 0,
                    "sex_conducteur1": "F",
                    "sex_conducteur2": "",
                    "anciennete_permis1": 34,
                    "anciennete_permis2": 0,
                    "anciennete_vehicule": 16.0,
                    "cylindre_vehicule": 1239,
                    "din_vehicule": 55,
                    "essence_vehicule": "Gasoline",
                    "marque_vehicule": "RENAULT",
                    "modele_vehicule": "CLIO",
                    "debut_vente_vehicule": 16,
                    "fin_vente_vehicule": 15,
                    "vitesse_vehicule": 150,
                    "type_vehicule": "Tourism",
                    "prix_vehicule": 10321,
                    "poids_vehicule": 830
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Persisted quote detail for the current customer context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "The required X-Client-ID header is missing or invalid."
          },
          "503": {
            "description": "Quote persistence is unavailable because PostgreSQL is not ready."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "List persisted quotes",
        "description": "Returns the quote history associated with the current browser context or authenticated account, ordered from the most recent to the oldest quote.",
        "operationId": "list_quotes",
        "parameters": [
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Persisted quote history for the current customer context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteListResponse"
                }
              }
            }
          },
          "400": {
            "description": "The required X-Client-ID header is missing or invalid."
          },
          "503": {
            "description": "Quote persistence is unavailable because PostgreSQL is not ready."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{quote_id}": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "Get a persisted quote",
        "description": "Returns one persisted quote when it belongs to the current browser context, the current account, or when it is accessed by an administrator.",
        "operationId": "get_quote",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote Id"
            }
          },
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Persisted quote detail for the current customer context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResponse"
                }
              }
            }
          },
          "400": {
            "description": "The required X-Client-ID header is missing or invalid."
          },
          "404": {
            "description": "The quote does not exist for the current customer context."
          },
          "503": {
            "description": "Quote persistence is unavailable because PostgreSQL is not ready."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quotes/{quote_id}/report.pdf": {
      "get": {
        "tags": [
          "quotes"
        ],
        "summary": "Generate the PDF report for one quote",
        "operationId": "download_quote_report",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote Id"
            }
          },
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/pdf": {}
            }
          },
          "404": {
            "description": "Quote not found."
          },
          "503": {
            "description": "Quote report generation is unavailable."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/quotes/{quote_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete a persisted quote",
        "operationId": "admin_delete_quote",
        "parameters": [
          {
            "name": "quote_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Quote Id"
            }
          },
          {
            "name": "X-Session-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Quote not found."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NumericValue": {
        "anyOf": [
          {
            "type": "integer"
          },
          {
            "type": "number"
          }
        ]
      },
      "PostalCodeValue": {
        "anyOf": [
          {
            "type": "integer"
          },
          {
            "type": "string"
          }
        ]
      },
...
      "PredictionInput": {
        "properties": {
          "index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index",
            "description": "Optional correlation identifier echoed back in prediction responses."
          },
          "bonus": {
            "$ref": "#/components/schemas/NumericValue",
            "description": "Driver bonus-malus coefficient."
          },
          "type_contrat": {
            "type": "string",
            "title": "Type Contrat",
            "description": "Insurance contract type."
          },
          "duree_contrat": {
            "type": "integer",
            "title": "Duree Contrat",
            "description": "Contract duration."
          },
          "anciennete_info": {
            "type": "integer",
            "title": "Anciennete Info",
            "description": "Years of customer history available."
          },
          "freq_paiement": {
            "type": "string",
            "title": "Freq Paiement",
            "description": "Premium payment frequency."
          },
          "paiement": {
            "type": "string",
            "title": "Paiement",
            "description": "Whether the premium is currently paid."
          },
          "utilisation": {
            "type": "string",
            "title": "Utilisation",
            "description": "Vehicle usage profile."
          },
          "code_postal": {
            "$ref": "#/components/schemas/PostalCodeValue",
            "description": "Postal code of the insured risk."
          },
          "conducteur2": {
            "type": "string",
            "title": "Conducteur2",
            "description": "Whether a secondary driver is declared."
          },
          "age_conducteur1": {
            "type": "integer",
            "title": "Age Conducteur1",
            "description": "Age of the primary driver."
          },
          "age_conducteur2": {
            "type": "integer",
            "title": "Age Conducteur2",
            "description": "Age of the secondary driver, if any."
          },
          "sex_conducteur1": {
            "type": "string",
            "title": "Sex Conducteur1",
            "description": "Sex of the primary driver."
          },
          "sex_conducteur2": {
            "type": "string",
            "title": "Sex Conducteur2",
            "description": "Sex of the secondary driver."
          },
          "anciennete_permis1": {
            "type": "integer",
            "title": "Anciennete Permis1",
            "description": "Driving licence seniority for driver 1."
          },
          "anciennete_permis2": {
            "type": "integer",
            "title": "Anciennete Permis2",
            "description": "Driving licence seniority for driver 2."
          },
          "anciennete_vehicule": {
            "$ref": "#/components/schemas/NumericValue",
            "description": "Vehicle age."
          },
          "cylindre_vehicule": {
            "type": "integer",
            "title": "Cylindre Vehicule",
            "description": "Vehicle engine displacement."
          },
          "din_vehicule": {
            "type": "integer",
            "title": "Din Vehicule",
            "description": "Vehicle power in DIN."
          },
          "essence_vehicule": {
            "type": "string",
            "title": "Essence Vehicule",
            "description": "Vehicle fuel type."
          },
          "marque_vehicule": {
            "type": "string",
            "title": "Marque Vehicule",
            "description": "Vehicle brand."
          },
          "modele_vehicule": {
            "type": "string",
            "title": "Modele Vehicule",
            "description": "Vehicle model."
          },
          "debut_vente_vehicule": {
            "type": "integer",
            "title": "Debut Vente Vehicule",
            "description": "Vehicle sale start indicator."
          },
          "fin_vente_vehicule": {
            "type": "integer",
            "title": "Fin Vente Vehicule",
            "description": "Vehicle sale end indicator."
          },
          "vitesse_vehicule": {
            "type": "integer",
            "title": "Vitesse Vehicule",
            "description": "Vehicle top speed."
          },
          "type_vehicule": {
            "type": "string",
            "title": "Type Vehicule",
            "description": "Vehicle category."
          },
          "prix_vehicule": {
            "type": "integer",
            "title": "Prix Vehicule",
            "description": "Vehicle price."
          },
          "poids_vehicule": {
            "type": "integer",
            "title": "Poids Vehicule",
            "description": "Vehicle weight."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "bonus",
          "type_contrat",
          "duree_contrat",
          "anciennete_info",
          "freq_paiement",
          "paiement",
          "utilisation",
          "code_postal",
          "conducteur2",
          "age_conducteur1",
          "age_conducteur2",
          "sex_conducteur1",
          "sex_conducteur2",
          "anciennete_permis1",
          "anciennete_permis2",
          "anciennete_vehicule",
          "cylindre_vehicule",
          "din_vehicule",
          "essence_vehicule",
          "marque_vehicule",
          "modele_vehicule",
          "debut_vente_vehicule",
          "fin_vente_vehicule",
          "vitesse_vehicule",
          "type_vehicule",
          "prix_vehicule",
          "poids_vehicule"
        ],
        "title": "InsurancePricingRecord",
        "examples": [
          {
            "age_conducteur1": 66,
            "age_conducteur2": 0,
            "anciennete_info": 1,
            "anciennete_permis1": 34,
            "anciennete_permis2": 0,
            "anciennete_vehicule": 16.0,
            "bonus": 0.58,
            "code_postal": 28388,
            "conducteur2": "No",
            "cylindre_vehicule": 1239,
            "debut_vente_vehicule": 16,
            "din_vehicule": 55,
            "duree_contrat": 1,
            "essence_vehicule": "Gasoline",
            "fin_vente_vehicule": 15,
            "freq_paiement": "Yearly",
            "index": 50000,
            "marque_vehicule": "RENAULT",
            "modele_vehicule": "CLIO",
            "paiement": "No",
            "poids_vehicule": 830,
            "prix_vehicule": 10321,
            "sex_conducteur1": "F",
            "sex_conducteur2": "",
            "type_contrat": "Maxi",
            "type_vehicule": "Tourism",
            "utilisation": "Retired",
            "vitesse_vehicule": 150
          }
        ]
      },
...
      "QuoteListResponse": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Number of quotes returned for the current client."
          },
          "quotes": {
            "items": {
              "$ref": "#/components/schemas/QuoteSummaryResponse"
            },
            "type": "array",
            "title": "Quotes",
            "description": "Quote history ordered from the most recent to the oldest quote."
          }
        },
        "type": "object",
        "required": [
          "count",
          "quotes"
        ],
        "title": "QuoteListResponse",
        "examples": [
          {
            "count": 1,
            "quotes": [
              {
                "created_at_utc": "2026-03-23T10:15:00+00:00",
                "id": "30bc53c8-5112-4a49-a204-5e8de5e7dcc5",
                "index": 50000,
                "marque_vehicule": "RENAULT",
                "modele_vehicule": "CLIO",
                "prime_prediction": 201.25504,
                "run_id": "base_v2_catboost_two_part_tweedie_1.3_train_smoke_42_classic_none_none",
                "type_contrat": "Maxi"
              }
            ]
          }
        ]
      },
      "QuoteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identifier of the persisted quote."
          },
          "created_at_utc": {
            "type": "string",
            "format": "date-time",
            "title": "Created At Utc",
            "description": "UTC timestamp at which the quote was created."
          },
          "run_id": {
            "type": "string",
            "title": "Run Id",
            "description": "Model bundle identifier used to compute the quote."
          },
          "input_payload": {
            "$ref": "#/components/schemas/PredictionInput",
            "description": "Original input payload used for the quote."
          },
          "result": {
            "$ref": "#/components/schemas/QuoteResultResponse",
            "description": "Prediction outputs returned for the quote."
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at_utc",
          "run_id",
          "input_payload",
          "result"
        ],
        "title": "QuoteResponse",
        "examples": [
          {
            "created_at_utc": "2026-03-23T10:15:00+00:00",
            "id": "30bc53c8-5112-4a49-a204-5e8de5e7dcc5",
            "input_payload": {
              "age_conducteur1": 66,
              "age_conducteur2": 0,
              "anciennete_info": 1,
              "anciennete_permis1": 34,
              "anciennete_permis2": 0,
              "anciennete_vehicule": 16.0,
              "bonus": 0.58,
              "code_postal": 28388,
              "conducteur2": "No",
              "cylindre_vehicule": 1239,
              "debut_vente_vehicule": 16,
              "din_vehicule": 55,
              "duree_contrat": 1,
              "essence_vehicule": "Gasoline",
              "fin_vente_vehicule": 15,
              "freq_paiement": "Yearly",
              "index": 50000,
              "marque_vehicule": "RENAULT",
              "modele_vehicule": "CLIO",
              "paiement": "No",
              "poids_vehicule": 830,
              "prix_vehicule": 10321,
              "sex_conducteur1": "F",
              "sex_conducteur2": "",
              "type_contrat": "Maxi",
              "type_vehicule": "Tourism",
              "utilisation": "Retired",
              "vitesse_vehicule": 150
            },
            "result": {
              "frequency_prediction": 0.0835,
              "index": 50000,
              "prime_prediction": 201.25504,
              "severity_prediction": 2410.24
            },
            "run_id": "base_v2_catboost_two_part_tweedie_1.3_train_smoke_42_classic_none_none"
          }
        ]
      },
      "QuoteResultResponse": {
        "properties": {
          "index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index",
            "description": "Input correlation identifier, returned when it was provided by the caller."
          },
          "frequency_prediction": {
            "type": "number",
            "title": "Frequency Prediction",
            "description": "Predicted claim frequency component."
          },
          "severity_prediction": {
            "type": "number",
            "title": "Severity Prediction",
            "description": "Predicted claim severity component."
          },
          "prime_prediction": {
            "type": "number",
            "title": "Prime Prediction",
            "description": "Final premium prediction for the quote."
          }
        },
        "type": "object",
        "required": [
          "frequency_prediction",
          "severity_prediction",
          "prime_prediction"
        ],
        "title": "QuoteResultResponse",
        "examples": [
          {
            "frequency_prediction": 0.0835,
            "index": 50000,
            "prime_prediction": 201.25504,
            "severity_prediction": 2410.24
          }
        ]
      },
...
    }
  },
...
}